% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETA.R
\name{BETA}
\alias{BETA}
\title{Evaluating generalized Beta distribution for use in the Rcompas}
\usage{
BETA(A0, M, R, ALPHA, GAMMA, X)
}
\arguments{
\item{A0}{A value.}

\item{M}{A vector of m values, as many as their is gradients.}

\item{R}{A vector of r values, as many as their is gradients.}

\item{ALPHA}{A vector of 'alpha' values, as many as their is gradients.}

\item{GAMMA}{A vector of 'gamma' values, as many as their is gradients.}

\item{X}{A matrix of gradients.}
}
\value{
A vector of values
}
\description{
Evaluates values of beta distribution with parameters provided by the user.
}
\author{
Marie-Hélène Ouellette
}
\keyword{datagen}
