% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DETER.R
\name{DETER}
\alias{DETER}
\title{Simulation of the deterministic effect}
\usage{
DETER(deter, height, varnor, nx, ny, centroide)
}
\arguments{
\item{deter}{Type of deterministic structure in resource variable ;(0), no deterministic structure, only N(0,varnor) is considered; (1), gradient from top (low values) to bottom (high values) + N(0,varnor); (2), gradients in two directions + N(0,varnor); (3), one big patch in center of the field + N(0,varnor); (4), 4 waves + N(0,varnor); (5), two zones separated by a sharp step + N(0,varnor).}

\item{height}{maximum value of the deterministic structure.}

\item{varnor}{variance of normal error.}

\item{nx}{Number of pixels on the x axis of the simulated surface.}

\item{ny}{Number of pixels on the y axis of the simulated surface.}

\item{centroide}{Vector of length 2 giving the position of the center of the multinormal distribution when using \code{deter}=3. Default is 0, putting the center of the multinormal distribution in the middle of the grid.}
}
\value{
A vector E containing the environment with deterministic structure
}
\description{
Simulation of the deterministic effect that structures species distribution
}
\references{
Legendre, P., M. R. T. Dale, M.-J. Fortin, J. Gurevitch, M. Hohn and D. Myers. 2002. The consequences of spatial structure for the design and analysis of ecological field surveys. \emph{Ecography} \strong{25}: 601-615.

Legendre, P., M. R. T. Dale, M.-J. Fortin, P. Casgrain and J. Gurevitch.  2004. Effects of spatial structures on the results of field experiments. \emph{Ecology} \strong{85} : 3202-3214.
}
\author{
Marie-Hélène Ouellette
}
\keyword{datagen}
