% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIST.R
\name{DIST}
\alias{DIST}
\title{Generating a value from a specified distribution for use in the Rcompas}
\usage{
DIST(name, n, ...)
}
\arguments{
\item{name}{One of 'unifrandom', 'lognormal', 'lograndom', 'exponential', 'gamma' and 'Poisson' distributions.}

\item{n}{The number of values to be generated.}

\item{\dots}{other arguments to be passed to generating functions.}
}
\value{
A vector of \code{n} values generated from the distribution specified by the user.
}
\description{
From which distribution should we take numbers (name), and how many (n)?
}
\author{
Marie-Hélène Ouellette
}
\keyword{Poisson'}
\keyword{datagen}
\keyword{exponential'}
\keyword{gamma'}
\keyword{lognormal'}
\keyword{lograndom'}
\keyword{normal'}
