% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUAL_NOISE.R
\name{QUAL_NOISE}
\alias{QUAL_NOISE}
\title{Add qualitative noise to a species simulated in Rcompas}
\usage{
QUAL_NOISE(SPi, Mi, Ri, ALPHAi, GAMMAi, X)
}
\arguments{
\item{SPi}{Piece of a BETAA list that contains a number of parameters that will be used later on to evaluate species abundances from the Rcompas function.}

\item{Mi}{m parameter for the specie.}

\item{Ri}{r parameter for the specie.}

\item{ALPHAi}{alpha parameter for the specie.}

\item{GAMMAi}{gamma parameter for the specie.}

\item{X}{the gradients on which the specie are simulated.}
}
\value{
the abundances of a species
}
\description{
Add qualitative noise to a species simulated in Rcompas, based on its beta parameters.
}
\author{
Marie-Hélène Ouellette
}
\keyword{datagen}
