% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsimSSDCOMPAS-package.R
\docType{package}
\name{RsimSSDCOMPAS-package}
\alias{RsimSSDCOMPAS-package}
\title{Simulation of ecological data}
\description{
Simulation of ecological data
}
\details{
\tabular{ll}{
Package: \tab RsimSSDCOMPAS\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2018-01-20\cr
License: \tab GPL (>= 2)\cr
}

This package encompasses various strategies to simulate ecological (species
and explanatory variables) data.
}
\author{
F. Guillaume Blanchet \email{guillaume.blanchet@usherbrooke.ca}
}
\keyword{package}
