% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPartial.R
\name{SimPartial}
\alias{SimPartial}
\title{Simulation of a covariate on top of a deterninistic effect}
\usage{
SimPartial(SP, ENV, type = "LIN", parms, corr)
}
\arguments{
\item{SP}{Species matrix to be modified.}

\item{ENV}{Environmental variable matrix to which W is linked.}

\item{type}{Type of covariate effect, either 'LIN' for linear, 'FACT' for factor (can be used for the binary case) or 'NORM' for normal density.
\itemize{
\item{LIN}{List with the following arguments in order : a vector containing the mean and sd of W, a vector giving the beta parameter of the regression : one for each species.}
\item{FACT}{In order the number of repetition of each states, and their effect (for example, 3 states, rep 10 times each on 30 sites, and their respective effect 3,5,8 would give the vector : c(3,10,10.10,3,5,8)}
\item{NORM}{List with the following arguments in order : a vector containing the mean and sd of W, a vector containing the maximum height of the effect of W : one for each species.}
}}

\item{parms}{Vector of parameters for the relationship between species and W. The number of elements and their meaning changes in function of the relationship between W and species.}

\item{corr}{Correlation matrix between E and W compeled by user. Must be square and symetric. Will change for FACT with the transformation of W.}
}
\value{
A list containing the new species data (SPW) and the resulting covariate (W).
}
\description{
Simulation of a covariate W that has an effect on the species distribution on top of a deterninistic effect.
}
\details{
The relationship between E and W is simulated using the Cholesky decomposition of the correlation matrix and the scaled E matrix both specified by the user. First, we use the function solve to get the inverse of the Cholesky decomposition of the variance matrix of the combined matrix with E and W (a random standardize normal vector). We multiply the combined matrix E, W with the resulting matrix. This insures we get EXACTLY the correlation intended. To conclude the simulation, we multiply the previoulsy obtained matrix with the Cholesky decomposition of the correlation matrix. The modified W vector now has the requested correlation with the original E matrix. Once this is establish, the effect on the species is simulated in different ways depending on the user's specification of type.
}
\author{
Marie-Hélène Ouellette
}
\keyword{datagen}
